refactor: modified the code and resolved the lint issue
