+++ cmd/git_test.go
@@ -44,29 +44,6 @@ func Test_GitCommand(t *testing.T) {
 		got := string(out)
 		assert.Contains(t, got, expected, "expected \"%s\" got \"%s\"", expected, got)
 	})
-	t.Run("Checking valid data", func(t *testing.T) {
-		outputBuffer := bytes.NewBufferString("")
-		rootCmd.SetOut(outputBuffer)
-		expected := "d"
-		rootCmd.SetArgs([]string{"git", expected})
-		err := gitCmd.Execute()
-		if err != nil {
-			assert.Error(t, err)
-		}
-		out, err := ioutil.ReadAll(outputBuffer)
-		if err != nil {
-			assert.Error(t, err)
-		}
-		got := string(out)
-		assert.Contains(t, got, "help", "expected \"%s\" got \"%s\"", "help", got)
-	})
-	t.Run("checking valid command", func(t *testing.T) {
-		outputBuffer := bytes.NewBufferString("")
-		rootCmd.SetOut(outputBuffer)
-		rootCmd.SetArgs([]string{"gi"})
-		err := gitCmd.Execute()
-		assert.Error(t, err)
-	})
 	t.Run("checking valid command", func(t *testing.T) {
 		outputBuffer := bytes.NewBufferString("")
 		rootCmd.SetOut(outputBuffer)
